chore : consolidate files from gg browser extension
