--- apps/browser-extension/services/zodSchema.ts
@@ -1,4 +1,0 @@
-import { isAddress } from "ethers/lib/utils";
-import { z } from "zod";
-
-export const Address = z.custom<string>(isAddress, "Invalid Address");
